get the stochastic MLP working
it works but needs a higher learning rate than the deterministic network policy.

Anyway, the thing that fixed this was an issue with the StochasticTanhGaussian network where there was an inappropriate sum over the logprob which essentially made the logprob regularization really strong and made all the gradients unnaturally large.